requirements.txt
numpy==2.2.6
pandas==2.2.3
Pillow==11.2.1
streamlit==1.45.1
tensorflow==2.19.0
